[http_service] Add some extra tests for the '/<model>/predict' endpoints
Fixes #1361
