--- http_service/tests/test_app.py
@@ -1,89 +1,0 @@
-# -*- coding: utf-8 -*-
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this file,
-# You can obtain one at http://mozilla.org/MPL/2.0/.
-
-import json
-
-from bugbug_http.app import API_TOKEN
-
-
-def test_empty_batch(client):
-    """Start with a blank database."""
-
-    rv = client.post(
-        "/component/predict/batch",
-        data=json.dumps({"bugs": []}),
-        headers={API_TOKEN: "test"},
-    )
-
-    assert rv.status_code == 400
-    assert rv.json == {"errors": {"bugs": ["min length is 1"]}}
-
-
-def test_too_big_batch(client):
-    """Start with a blank database."""
-
-    bugs = list(range(1001))
-
-    rv = client.post(
-        "/component/predict/batch",
-        data=json.dumps({"bugs": bugs}),
-        headers={API_TOKEN: "test"},
-    )
-
-    assert rv.status_code == 400
-    assert rv.json == {"errors": {"bugs": ["max length is 1000"]}}
-
-
-def test_non_int_batch(client):
-    """Start with a blank database."""
-
-    bugs = ["1", "2", "3"]
-
-    rv = client.post(
-        "/component/predict/batch",
-        data=json.dumps({"bugs": bugs}),
-        headers={API_TOKEN: "test"},
-    )
-
-    assert rv.status_code == 400
-    assert rv.json == {
-        "errors": {
-            "bugs": [
-                {
-                    "0": ["must be of integer type"],
-                    "1": ["must be of integer type"],
-                    "2": ["must be of integer type"],
-                }
-            ]
-        }
-    }
-
-
-def test_unknown_model(client):
-    """Start with  a blank database,"""
-    bugs = [1, 2, 3]
-    unknown_model = "unknown_model"
-
-    rv = client.post(
-        f"/{unknown_model}/predict/batch",
-        data=json.dumps({"bugs": bugs}),
-        headers={API_TOKEN: "test"},
-    )
-
-    assert rv.status_code == 404
-    assert rv.json == {"error": f"Model {unknown_model} doesn't exist"}
-
-
-def test_no_api_key(client):
-    """Start with an empty database,"""
-
-    bugs = [1, 2, 3]
-
-    rv = client.post(
-        "/component/predict/batch", data=json.dumps({"bugs": bugs}), headers={},
-    )
-
-    assert rv.status_code == 401
-    assert rv.json == {"message": "Error, missing X-API-KEY"}
